+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [5.1.0-RC-12]
 ### Changed
 - Nested Step finish refactoring
 ### Added